requirements.txt
pytest==7.1.3
allure-pytest==2.12.0
selenium==4.5.0
webdriver-manager==3.8.3
pytest-xdist==2.5.0
sqlalchemy==1.4.44
pymysql==1.0.2
cryptography==38.0.3
